A few tweaks and an update to the ReadMe.
First of all, I wanted to try and make the help message a little
more precise. Second, I wanted to add the / to the cp command
to try and prevent situations whre the drive would get copied
onto a file called current if the directory didn't exist.

Updated the ReadMe to be explicit about what's going on.